requirements_server.txt
flask
flask-cors
paddleocr
opencv-python
numpy
pillow
pymupdf
